problems with flickering background

+++ include/camera.h
@@ -17,7 +17,7 @@ public:
 	int y();
 
 private:
-	int x_, y_;
+	float x_, y_;
 	int width_, height_;
 	int xMin_, yMin_, xMax_, yMax_;
 	Sprite* focus_;
